fix(doctor): compare statfs magics portably so the musl images build
The offline doctor launcher, worker, collector and provisioner are only
distributable as x86_64-unknown-linux-musl: scripts/package-doctor-release.sh
admits no other target. They had never been compiled for it. glibc declares
statfs.f_type signed and musl unsigned, so five comparisons against
libc::TMPFS_MAGIC, CGROUP2_SUPER_MAGIC and PROC_SUPER_MAGIC failed to type
check and the release images could not be built at all.

Compare the 64-bit kernel field's bits instead of the libc declaration's
signedness. The admitted magics are unchanged, and the source-layout tripwire's
required substrings are preserved.

+++ crates/semaprax-native-rust-interop-platform-sys/src/doctor/offline_input/linux.rs
@@ -116,7 +116,9 @@ pub(super) fn snapshot(
     } else {
         filesystem
     };
-    if filesystem.f_type != libc::TMPFS_MAGIC {
+    // glibc declares statfs.f_type signed and musl unsigned; the kernel field
+    // is the same 64-bit value, so compare its bits, not the libc signedness.
+    if filesystem.f_type as u64 != libc::TMPFS_MAGIC as u64 {
         return Err(Error::Invalid);
     }
